Fixes for simc changes since last update.

* Changes to how the scaling information is pulled through to spell effects instead of spells.
* Added missing fields.
* Updated the -9 scaling implementation and added tests.
* Fixed broken tests.

+++ SimcProfileParser/Model/RawData/SimcRawRandomPropData.cs
@@ -1,10 +1,13 @@
 ﻿namespace SimcProfileParser.Model.RawData
 {
+    /// <summary>
+    /// rand_prop_points.hpp - random_prop_data_t
+    /// </summary>
     class SimcRawRandomPropData
     {
         public uint ItemLevel { get; set; }
-        public uint DamageReplaceStat { get; set; }
-        public uint DamageSecondary { get; set; }
+        public double DamageReplaceStat { get; set; }
+        public double DamageSecondary { get; set; }
         public float[] Epic { get; set; }
         public float[] Rare { get; set; }
         public float[] Uncommon { get; set; }
